refactor: clean up service exports by removing unused services
