+++ crates/forge_app/src/lib.rs
@@ -3,7 +3,4 @@ mod schema;
 mod service;
 mod sqlite;
 
-pub use service::{
-    APIService, ChatService, ConversationHistory, EnvironmentService, FileReadService, Service,
-    SuggestionService, TitleService, UIService,
-};
+pub use service::{APIService, Service};
