+++ crates/forge_app/src/service/mod.rs
@@ -13,13 +13,6 @@ mod user_prompt;
 mod workflow_title;
 
 pub use api::APIService;
-pub use chat::{ChatService, ConversationHistory};
-pub use env::EnvironmentService;
-pub use file_read::FileReadService;
-pub use suggestion::{File, SuggestionService};
-pub use ui::UIService;
-pub use workflow_title::TitleService;
-
 pub struct Service;
 
 #[async_trait::async_trait]
